services/factorial/requirements.txt
googleapis-common-protos==1.52.0
grpcio==1.34.0
grpcio-status==1.34.0
grpcio-tools==1.34.0
protobuf==3.14.0
six==1.15.0
